Move more tests to HoverflyTest
