+++ dockstore-integration-testing/src/test/java/io/dockstore/webservice/TokenResourceIT.java
@@ -40,6 +40,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import io.dockstore.client.cli.BaseIT.TestStatus;
 import io.dockstore.common.CommonTestUtilities;
+import io.dockstore.common.HoverflyTest;
 import io.dockstore.common.MuteForSuccessfulTests;
 import io.dockstore.common.NonConfidentialTest;
 import io.dockstore.common.TestingPostgres;
@@ -91,6 +92,7 @@ import uk.org.webcompere.systemstubs.stream.SystemOut;
 @ExtendWith(HoverflyExtension.class)
 @HoverflyCore(mode = HoverflyMode.SIMULATE)
 @Tag(NonConfidentialTest.NAME)
+@Tag(HoverflyTest.NAME)
 public class TokenResourceIT {
     // This is not from Hoverfly, it's actually in the starting database
     public static final String GITHUB_ACCOUNT_USERNAME = "potato";
