+++ dockstore-integration-testing/src/test/java/io/dockstore/webservice/WebhookIT.java
@@ -2,7 +2,6 @@
 package io.dockstore.webservice;
 
 import static io.dockstore.client.cli.WorkflowIT.DOCKSTORE_TEST_USER_2_HELLO_DOCKSTORE_NAME;
-import static io.dockstore.common.Hoverfly.ORCID_SIMULATION_SOURCE;
 import static io.dockstore.webservice.Constants.DOCKSTORE_YML_PATH;
 import static io.dockstore.webservice.helpers.GitHubAppHelper.LAMBDA_ERROR;
 import static io.dockstore.webservice.helpers.GitHubAppHelper.handleGitHubBranchDeletion;
@@ -45,7 +44,6 @@ import io.dockstore.openapi.client.api.WorkflowsApi;
 import io.dockstore.openapi.client.model.Collection;
 import io.dockstore.openapi.client.model.Entry;
 import io.dockstore.openapi.client.model.LambdaEvent;
-import io.dockstore.openapi.client.model.OrcidAuthorInformation;
 import io.dockstore.openapi.client.model.Organization;
 import io.dockstore.openapi.client.model.PublishRequest;
 import io.dockstore.openapi.client.model.SourceFile;
@@ -67,8 +65,6 @@ import io.dockstore.webservice.jdbi.NotebookDAO;
 import io.dockstore.webservice.jdbi.WorkflowVersionDAO;
 import io.dockstore.webservice.languages.WDLHandler;
 import io.dropwizard.client.JerseyClientBuilder;
-import io.specto.hoverfly.junit.core.Hoverfly;
-import io.specto.hoverfly.junit.core.HoverflyMode;
 import jakarta.ws.rs.client.Client;
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -729,45 +725,6 @@ class WebhookIT extends BaseIT {
         workflow = getFoobarWorkflowDockstoreTesting(client);
         assertEquals("manualTopic", workflow.getDefaultVersion());
         assertEquals(2, workflow.getWorkflowVersions().size());
-    }
-
-    /**
-     * This test relies on Hoverfly to simulate responses from the ORCID API.
-     * In the simulation, the responses are crafted for an ORCID author with ID 0000-0002-6130-1021.
-     * ORCID authors with other IDs are considered "not found" by the simulation.
-     */
-    @Test
-    void testGetWorkflowVersionOrcidAuthors() {
-        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowsApi = new WorkflowsApi(webClient);
-        ApiClient anonymousWebClient = getAnonymousOpenAPIWebClient();
-        WorkflowsApi anonymousWorkflowsApi = new WorkflowsApi(anonymousWebClient);
-        String wdlWorkflowRepoPath = String.format("github.com/%s/%s", DockstoreTestUser2.TEST_AUTHORS, "foobar");
-
-        // Workflows containing 1 descriptor author and multiple .dockstore.yml authors.
-        // If the .dockstore.yml specifies an author, then only the .dockstore.yml's authors should be saved
-        handleGitHubRelease(workflowsApi, DockstoreTestUser2.TEST_AUTHORS, "refs/heads/main", USER_2_USERNAME);
-        // WDL workflow
-        Workflow workflow = workflowsApi.getWorkflowByPath(wdlWorkflowRepoPath, WorkflowSubClass.BIOWORKFLOW, "versions,authors");
-        WorkflowVersion version = workflow.getWorkflowVersions().stream().filter(v -> v.getName().equals("main")).findFirst().get();
-        assertEquals(2, version.getAuthors().size());
-        assertEquals(2, version.getOrcidAuthors().size());
-
-        // Hoverfly is not used as a class rule here because for some reason it's trying to intercept GitHub in both spy and simulation mode
-        try (Hoverfly hoverfly = new Hoverfly(HoverflyMode.SIMULATE)) {
-            hoverfly.start();
-            hoverfly.simulate(ORCID_SIMULATION_SOURCE);
-            List<OrcidAuthorInformation> orcidAuthorInfo = workflowsApi.getWorkflowVersionOrcidAuthors(workflow.getId(), version.getId());
-            assertEquals(1, orcidAuthorInfo.size()); // There's 1 OrcidAuthorInfo instead of 2 because only 1 ORCID ID from the version exists on ORCID
-
-            // Publish workflow
-            PublishRequest publishRequest = CommonTestUtilities.createOpenAPIPublishRequest(true);
-            workflowsApi.publish1(workflow.getId(), publishRequest);
-
-            // Check that an unauthenticated user can get the workflow version ORCID authors of a published workflow
-            anonymousWorkflowsApi.getWorkflowVersionOrcidAuthors(workflow.getId(), version.getId());
-            assertEquals(1, orcidAuthorInfo.size());
-        }
     }
 
     /**
